[Refactor] Refactor project structure

+++ glTFRenderer/RHICore/RHICore/Private/RHIVKImpl/VKPipelineStateObject.cpp
@@ -98,9 +98,9 @@ bool VKGraphicsPipelineStateObject::InitPipelineStateObject(IRHIDevice& device,
     
     VkViewport viewport{};
     viewport.x = 0.0f;
-    viewport.y = scissor.extent.height;
-    viewport.width = (float) scissor.extent.width;
-    viewport.height = (float) -scissor.extent.height;
+    viewport.y = static_cast<float>(scissor.extent.height);
+    viewport.width = static_cast<float>(scissor.extent.width);
+    viewport.height = -static_cast<float>(scissor.extent.height);
     viewport.minDepth = 0.0f;
     viewport.maxDepth = 1.0f;
